gausium_legacy_connector/requirements.txt
# SPDX-FileCopyrightText: 2025 InOrbit, Inc.
#
# SPDX-License-Identifier: MIT

inorbit-connector==1.4.0
httpx~=0.28
pytz>=2022.7
ruamel.yaml>=0.18,<0.19
pydantic>=2.6,<3.0
pydantic-settings>=2.6,<3.0
python-dotenv==1
pillow==10.4.0
tenacity==9.1.2
